server/requirements.txt

fastapi
uvicorn
python-multipart
google-cloud-storage
google-cloud-vision
google-cloud-aiplatform
Pillow
